backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.3.0
python-multipart==0.0.6
aiofiles==24.1.0
httpx==0.25.2
tenacity==8.2.3
pyspellchecker==0.8.1 